requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
databases==0.9.0
fastapi==0.115.5
h11==0.14.0
httptools==0.6.4
idna==3.10
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.0
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
